Add support for python 3.4
The only incompatibility is the long integer type. Mixed int and long
calculations are automatic in python 2, and there is no long type in
python 3, so all mentions of "long" are removed. To test if an object is
an integer type, six.integer_types is used, which evaluates to (int,
long) in python 2 and (int) in python 3.